+++ Source/OnlineSubsystemAccelByte/Private/AsyncTasks/OnlineAsyncTaskAccelByteLogin.cpp
@@ -35,8 +35,14 @@ void FOnlineAsyncTaskAccelByteLogin::Initialize()
 	Super::Initialize();
 
 	AB_OSS_ASYNC_TASK_TRACE_BEGIN(TEXT("LocalUserNum: %d"), LocalUserNum);
-
-	ApiClient = MakeShared<AccelByte::FApiClient, ESPMode::ThreadSafe>();
+	if (Subsystem->IsMultipleLocalUsersEnabled())
+	{
+		ApiClient = MakeShared<AccelByte::FApiClient, ESPMode::ThreadSafe>();
+	}
+	else
+	{
+		ApiClient = FMultiRegistry::GetApiClient();
+	}
 	ApiClient->CredentialsRef->SetClientCredentials(FRegistry::Settings.ClientId, FRegistry::Settings.ClientSecret);
 
 	// If all members of the AccountCredentials struct are blank, then this should be treated as a pass through login to
@@ -74,11 +80,7 @@ void FOnlineAsyncTaskAccelByteLogin::Finalize()
 {
 	AB_OSS_ASYNC_TASK_TRACE_BEGIN(TEXT("bWasSuccessful: %s"), LOG_BOOL_FORMAT(bWasSuccessful));
 
-	const FOnlineIdentityAccelBytePtr IdentityInterface = FOnlineIdentityAccelByte::Get();
-	if (IdentityInterface.IsValid())
-	{
-		IdentityInterface->SetLocalUserNumCached(LocalUserNum);
-	}
+	Subsystem->SetLocalUserNumCached(LocalUserNum);
 
 	AB_OSS_ASYNC_TASK_TRACE_END(TEXT(""));
 }
@@ -87,13 +89,15 @@ void FOnlineAsyncTaskAccelByteLogin::TriggerDelegates()
 {
 	AB_OSS_ASYNC_TASK_TRACE_BEGIN(TEXT("bWasSuccessful: %s"), LOG_BOOL_FORMAT(bWasSuccessful));
 
-	const FOnlineIdentityAccelBytePtr IdentityInterface = FOnlineIdentityAccelByte::Get();
+	const FOnlineIdentityAccelBytePtr IdentityInterface = StaticCastSharedPtr<FOnlineIdentityAccelByte>(Subsystem->GetIdentityInterface());
+			
 	if (IdentityInterface.IsValid())
 	{
 		IdentityInterface->SetLoginStatus(LocalUserNum, bWasSuccessful ? ELoginStatus::LoggedIn : ELoginStatus::NotLoggedIn);
 		if (bWasSuccessful && !ApiClient->CredentialsRef->IsComply())
 		{
-			const FOnlineAgreementAccelBytePtr AgreementInterface = FOnlineAgreementAccelByte::Get();
+			const FOnlineAgreementAccelBytePtr AgreementInterface = StaticCastSharedPtr<FOnlineAgreementAccelByte>(Subsystem->GetAgreementInterface());
+			
 			AgreementInterface->TriggerOnUserNotCompliedDelegates(LocalUserNum);
 		}
 		else