+++ Source/OnlineSubsystemAccelByte/Private/AsyncTasks/OnlineAsyncTaskAccelByteRegisterPlayer.cpp
@@ -16,11 +16,7 @@ FOnlineAsyncTaskAccelByteRegisterPlayers::FOnlineAsyncTaskAccelByteRegisterPlaye
 	, bWasInvited(InBWasInvited)
 	, bIsSpectator(InBIsSpectator)
 {
-	FOnlineIdentityAccelBytePtr IdentityInterface = StaticCastSharedPtr<FOnlineIdentityAccelByte>(Subsystem->GetIdentityInterface());
-	if (IdentityInterface.IsValid())
-	{
-		LocalUserNum = IdentityInterface->GetLocalUserNumCached();
-	}
+	LocalUserNum = Subsystem->GetLocalUserNumCached();
 }
 
 void FOnlineAsyncTaskAccelByteRegisterPlayers::Initialize()