+++ Source/OnlineSubsystemAccelByte/Public/OnlineAgreementInterfaceAccelByte.h
@@ -62,16 +62,6 @@ public:
 
 	DEFINE_ONLINE_PLAYER_DELEGATE_TWO_PARAM(MAX_LOCAL_PLAYERS, OnAcceptAgreementPoliciesCompleted, bool, const FString&);
 
-	static FOnlineAgreementAccelBytePtr Get()
-	{
-		if (IOnlineSubsystem::DoesInstanceExist(ACCELBYTE_SUBSYSTEM))
-		{
-			const FOnlineSubsystemAccelByte* AccelByteSubsystem = static_cast<FOnlineSubsystemAccelByte*>(IOnlineSubsystem::Get(ACCELBYTE_SUBSYSTEM));
-			return AccelByteSubsystem ? StaticCastSharedPtr<FOnlineAgreementAccelByte>(AccelByteSubsystem->GetAgreementInterface()) : nullptr;
-		}
-		return nullptr;
-	}
-
 	/**
 	 * @brief Query user's eligibilities
 	 *